backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
groq==0.4.1
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0